requirements.txt

python-dotenv==1.0.0
langchain==0.3.27
langchain-community==0.3.31
sentence-transformers==5.1.1
faiss-cpu==1.12.0
langchain-openai==0.3.35
openai==2.4.0
pytest==8.4.2
fastapi==0.119.0
uvicorn==0.37.0
pydantic==2.12.2
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
gradio
numpy<2
awscli
boto3